setAssigned: Handle officers
If you assign crew to quartermaster or sailing master, then they're in charge of either the starboard crew or the larboard crew. As such, they need to be added to that team's crew, and they cannot be a member of the other team's crew. That also means that you can't be the other team's helm or lookout, and you can't be a quartermaster and a sailing master at the same time.