Remove timezone handling
The Caribbean is in the UTC now. Not really, but this is centuries ago, before there were timezones, before there was even a Gregorian calendar, so who cares? Times need to be accurate relative to one another, not relative to any actual, real-world timekeeping standard. So long as everything is UTC, it should be fine (right?).