CrewPanel: Drop your ship
But hey, did you know that picking up the adventure document is an async operation? Which meant switching getAdventure to async, which meant switching getCrawlState to async, which meant switching MOST OF THE EXISTING CODEBASE TO ASYNC. I'm honestly surprised that I can make a commit with 88 FILES CHANGED and not only do all the tests pass, it actually does the thing it's supposed to do.